Rename CTS1 to CIS1
