+++ nft-client/src/lib.rs
@@ -1,1 +1,1 @@
-pub mod cts1;
+pub mod cis1;
